Split streams.ts into streamGroups.ts, streamStreamers.ts, streamRestart.ts
Same qlty "high total complexity" fix as the earlier streamdeck.ts split:
adding the atomic removeStreamGroupAndStreamers plus the updateStreamGroup
boolean check pushed streams.ts's file-total complexity over the threshold.
Groups and streamers are separate concerns anyway, so split their routes
into their own files; the shared restart-serialization helper moves to
streamRestart.ts so both can use it. streams.ts is now just the GET view/
live-state routes plus composition.
